Fix add method in view
