+++ NotesApplication/ApplicationServiceCollectionExtensions.cs
@@ -1,0 +1,22 @@
+﻿using System.Reflection;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace NotesApplication;
+
+/// <summary>
+/// Class ApplicationServiceCollectionExtensions.
+/// </summary>
+public static class ApplicationServiceCollectionExtensions
+{
+    /// <summary>
+    /// Adds the application.
+    /// </summary>
+    /// <param name="services">The services.</param>
+    /// <returns>IServiceCollection.</returns>
+    public static IServiceCollection AddApplication(this IServiceCollection services)
+    {
+        services.AddMediatR(cfg => cfg.RegisterServicesFromAssembly(Assembly.GetExecutingAssembly()));
+
+        return services;
+    }
+}
