+++ NotesApplication/Data/Repository.cs
@@ -1,16 +1,34 @@
 ﻿using NotesPresistence;
 
-namespace NotesWebApi.Data;
+namespace NotesApplication.Data;
 
+/// <summary>
+/// Class Repository.
+/// Implements the <see cref="NotesApplication.Data.IRepository{TEntity}" />
+/// </summary>
+/// <typeparam name="TEntity">The type of the t entity.</typeparam>
+/// <seealso cref="NotesApplication.Data.IRepository{TEntity}" />
 public class Repository<TEntity> : IRepository<TEntity> where TEntity : class, new()
 {
+    /// <summary>
+    /// The database context
+    /// </summary>
     private readonly NotesDbContext _dbContext;
 
+    /// <summary>
+    /// Initializes a new instance of the <see cref="Repository{TEntity}"/> class.
+    /// </summary>
+    /// <param name="notesDbContext">The notes database context.</param>
     public Repository(NotesDbContext notesDbContext)
     {
         _dbContext = notesDbContext;
     }
 
+    /// <summary>
+    /// Gets all.
+    /// </summary>
+    /// <returns>IQueryable&lt;TEntity&gt;.</returns>
+    /// <exception cref="NotesApplication.Common.Exception">Couldn't retrieve entities</exception>
     public IQueryable<TEntity> GetAll()
     {
         try
@@ -24,6 +42,13 @@ public class Repository<TEntity> : IRepository<TEntity> where TEntity : class, n
         }
     }
 
+    /// <summary>
+    /// Add as an asynchronous operation.
+    /// </summary>
+    /// <param name="entity">The entity.</param>
+    /// <returns>A Task&lt;TEntity&gt; representing the asynchronous operation.</returns>
+    /// <exception cref="System.ArgumentNullException"></exception>
+    /// <exception cref="NotesApplication.Common.Exception"></exception>
     public async Task<TEntity> AddAsync(TEntity entity)
     {
         if (entity == null)
@@ -43,6 +68,13 @@ public class Repository<TEntity> : IRepository<TEntity> where TEntity : class, n
         }
     }
 
+    /// <summary>
+    /// Update as an asynchronous operation.
+    /// </summary>
+    /// <param name="entity">The entity.</param>
+    /// <returns>A Task&lt;TEntity&gt; representing the asynchronous operation.</returns>
+    /// <exception cref="System.ArgumentNullException"></exception>
+    /// <exception cref="NotesApplication.Common.Exception"></exception>
     public async Task<TEntity> UpdateAsync(TEntity entity)
     {
         if (entity == null)
@@ -62,6 +94,12 @@ public class Repository<TEntity> : IRepository<TEntity> where TEntity : class, n
             throw new Exception($"{nameof(entity)} could not be updated");
         }
     }
+    /// <summary>
+    /// Delete as an asynchronous operation.
+    /// </summary>
+    /// <param name="entity">The entity.</param>
+    /// <returns>A Task&lt;TEntity&gt; representing the asynchronous operation.</returns>
+    /// <exception cref="NotesApplication.Common.Exception"></exception>
     public async Task<TEntity> DeleteAsync(TEntity entity)
     {
         try
