+++ NotesApplication/Notes/Commands/DeleteCommand/DeleteNoteCommand.cs
@@ -1,0 +1,9 @@
+﻿using MediatR;
+
+namespace NotesApplication.Notes.Commands.DeleteCommand;
+
+public class DeleteNoteCommand : IRequest<Unit>
+{
+    public string UserId { get; set; }
+    public string Id { get; set; }
+}
