+++ NotesApplication/Notes/Commands/UpdateNote/UpdateNoteCommandHandler.cs
@@ -3,18 +3,7 @@ using Microsoft.EntityFrameworkCore;
 using NotesApplication.Common.Exception;
 using NotesPresistence;
 
-namespace NotesWebApi.Notes.Commands.UpdateNote;
-
-public class UpdateNoteCommand : IRequest<Unit>
-{
-    public string Id { get; set; }
-
-    public string UserId { get; set; }
-
-    public string? Title { get; set; }
-
-    public string? Details { get; set; }
-}
+namespace NotesApplication.Notes.Commands.UpdateNote;
 
 public class UpdateNoteCommandHandler : IRequestHandler<UpdateNoteCommand, Unit>
 {
