+++ NotesApplication/Notes/Queries/GetNoteDetails/GetNoteDatailsQuery.cs
@@ -1,6 +1,6 @@
 ﻿using MediatR;
 using NotesPresistence;
 
-namespace NotesWebApi.Notes.Queries.GetNoteDetails;
+namespace NotesApplication.Notes.Queries.GetNoteDetails;
 public record GetNoteDatailsQuery(string? UserId, string Id) :  IRequest<List<NoteDetailsVm>?>;
