+++ NotesApplication/Notes/Queries/GetNotesList/GetListNoteQuery.cs
@@ -1,6 +1,6 @@
 ﻿using MediatR;
 using NotesPresistence;
 
-namespace NotesWebApi.Notes.Queries.GetNotesList;
+namespace NotesApplication.Notes.Queries.GetNotesList;
 
 public record GetListNoteQuery(string? UserId) : IRequest<List<NoteLookUpDto>?>;
