+++ NotesApplication/Notes/Queries/GetNotesList/NoteListVm.cs
@@ -1,6 +1,6 @@
 ﻿using NotesPresistence;
 
-namespace NotesWebApi.Notes.Queries.GetNotesList;
+namespace NotesApplication.Notes.Queries.GetNotesList;
 
 public class NoteListVm
 {
