--- NotesWebApi/Data/IRepository.cs
@@ -1,10 +1,0 @@
-﻿namespace NotesWebApi.Data;
-
-public interface IRepository<TEntity> where TEntity : class, new()
-{
-    IQueryable<TEntity> GetAll();
-
-    Task<TEntity> AddAsync(TEntity entity);
-
-    Task<TEntity> UpdateAsync(TEntity entity);
-}
