--- NotesWebApi/Models/UpdateNoteDto.cs
@@ -1,21 +1,0 @@
-﻿using AutoMapper;
-using NotesPresistence.Mappings;
-using NotesWebApi.Notes.Commands.UpdateNote;
-
-namespace NotesWebApi.Models;
-
-public class UpdateNoteDto : IMapWith<UpdateNoteDto>
-{
-    public string Id { get; set; }
-    public string Title { get; set; }
-    public string Details { get; set; }
-
-    public void Mapping(Profile profile)
-    {
-        profile.CreateMap<UpdateNoteDto, UpdateNoteCommand>()
-            .ForMember(noteCommand => noteCommand.Title,
-                opt => opt.MapFrom(noteDto => noteDto.Title))
-            .ForMember(noteCommand => noteCommand.Details,
-                opt => opt.MapFrom(noteDto => noteDto.Details));
-    }
-}
