+++ NotesWebApi/Startup.cs
@@ -1,65 +1,65 @@
-﻿using System.Reflection;
-using NotesApplication;
+﻿using NotesApplication;
 using NotesPresistence;
 using NotesPresistence.Mapping;
-using NotesWebApi.Data;
+using System.Reflection;
+using NotesApplication.Data;
 
-namespace NotesWebApi
+namespace NotesWebApi;
+
+public class Startup
 {
-    public class Startup
+    public IConfiguration Configuration { get; set; }
+    public Startup(IConfiguration configuration)
     {
-        public IConfiguration Configuration { get; set; }
-        public Startup(IConfiguration configuration)
+        Configuration = configuration;
+    }
+
+    public void ConfigureServices(IServiceCollection services)
+    {
+        services.AddScoped(typeof(IRepository<>), typeof(Repository<>));
+        services.AddAutoMapper(config =>
         {
-            Configuration = configuration;
-        }
+            config.AddProfile(new AssemblyMappingProfile(Assembly.GetExecutingAssembly()));
+            config.AddProfile(new AssemblyMappingProfile(typeof(INotesDbContext).Assembly));
+        });
+        services.AddApplicaton();
+        services.AddPersistence(Configuration);
+        services.AddControllers();
 
-        public void ConfigureServices(IServiceCollection services)
+        services.AddApplication();
+        //services.AddMediatR(cfg => cfg.RegisterServicesFromAssembly(typeof(Startup).Assembly));
+
+        services.AddCors(options =>
         {
-            services.AddScoped(typeof(IRepository<>), typeof(Repository<>));
-            services.AddAutoMapper(config =>
+            options.AddPolicy("AllowAll", policy =>
             {
-                config.AddProfile(new AssemblyMappingProfile(Assembly.GetExecutingAssembly()));
-                config.AddProfile(new AssemblyMappingProfile(typeof(INotesDbContext).Assembly));
+                policy.AllowAnyHeader();
+                policy.AllowAnyMethod();
+                policy.AllowAnyOrigin();
             });
-            services.AddApplicaton();
-            services.AddPersistence(Configuration);
-            services.AddControllers();
+        });
+        services.AddSwaggerGen();
+    }
 
-            services.AddMediatR(cfg => cfg.RegisterServicesFromAssembly(typeof(Startup).Assembly));
-
-            services.AddCors(options =>
-            {
-                options.AddPolicy("AllowAll", policy =>
-                {
-                    policy.AllowAnyHeader();
-                    policy.AllowAnyMethod();
-                    policy.AllowAnyOrigin();
-                });
-            });
-            services.AddSwaggerGen();
-        }
-
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+    public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
+    {
+        if (env.IsDevelopment())
         {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-            }
-            app.UseSwagger();
-            app.UseSwaggerUI(config =>
-            {
-                config.RoutePrefix = string.Empty;
-                config.SwaggerEndpoint("swagger/v1/swagger.json", "Notes API");
-            });
-            app.UseRouting();
-            app.UseHttpsRedirection();
-            app.UseCors();
-
-            app.UseEndpoints(endpoints =>
-            {
-                endpoints.MapControllers();
-            });
+            app.UseDeveloperExceptionPage();
         }
+        app.UseSwagger();
+        app.UseSwaggerUI(config =>
+        {
+            config.RoutePrefix = string.Empty;
+            config.SwaggerEndpoint("swagger/v1/swagger.json", "Notes API");
+        });
+        app.UseRouting();
+        app.UseHttpsRedirection();
+        app.UseCors();
+
+        app.UseEndpoints(endpoints =>
+        {
+            endpoints.MapControllers();
+        });
     }
 }
