+++ WebClient/WebClient.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
     <TargetFramework>net7.0</TargetFramework>
@@ -11,6 +11,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\NotesApplication\NotesApplication.csproj" />
     <ProjectReference Include="..\NotesPresistence\NotesPresistence.csproj" />
   </ItemGroup>
