+++ WebClient/_Imports.razor
@@ -5,6 +5,7 @@
 @using Microsoft.JSInterop
 @using NotesPresistence
 @using System.ComponentModel
+@using System.ComponentModel.DataAnnotations
 @using WebClient
 @using WebClient.Components
 @using WebClient.Pages
